requirements.txt
# ── TradAI — Dependencias ─────────────────────────────────

# Base de datos
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
greenlet>=3.0

# Telegram
python-telegram-bot>=21.0

# Datos de mercado
yfinance>=0.2.30

# IA (OpenAI)
openai>=1.30

# Noticias
aiohttp>=3.9
feedparser>=6.0

# Scheduler
APScheduler>=3.10

# Configuración
python-dotenv>=1.0

# Utilidades
pandas>=2.0
numpy>=1.24
lxml>=5.0
html5lib>=1.1

# Calendario de festivos (opcional pero recomendado)
exchange_calendars>=4.5

# Dashboard web
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1

# Migraciones
alembic>=1.13

# Tests
pytest>=8.0
pytest-asyncio>=0.23
